Function file restruturing
