+++ .secrets.baseline
@@ -183,21 +183,21 @@
         "filename": "tests/io/conftest.py",
         "hashed_secret": "adb5fabe51f5b45e83fdd91b71c92156fec4a63e",
         "is_verified": false,
-        "line_number": 71
+        "line_number": 70
       },
       {
         "type": "Secret Keyword",
         "filename": "tests/io/conftest.py",
         "hashed_secret": "3c3b274d119ff5a5ec6c1e215c1cb794d9973ac1",
         "is_verified": false,
-        "line_number": 117
+        "line_number": 116
       },
       {
         "type": "Secret Keyword",
         "filename": "tests/io/conftest.py",
         "hashed_secret": "15dd2c9ccec914f1470b4dccb45789844e49cf70",
         "is_verified": false,
-        "line_number": 131
+        "line_number": 130
       }
     ],
     "tests/io/test_data_catalog.py": [
@@ -219,5 +219,5 @@
       }
     ]
   },
-  "generated_at": "2025-04-07T11:15:37Z"
+  "generated_at": "2025-06-14T15:25:52Z"
 }
