+++ tests/io/conftest.py
@@ -4,6 +4,7 @@ import pytest
 from kedro_datasets.pandas import CSVDataset
 
 from kedro.io import CachedDataset, Version
+from tests.testing_utils import dummy_dataframe
 
 
 @pytest.fixture
@@ -11,9 +12,7 @@ def dummy_numpy_array():
     return np.array([[1, 4, 5], [2, 5, 6]])
 
 
-@pytest.fixture
-def dummy_dataframe():
-    return pd.DataFrame({"col1": [1, 2], "col2": [4, 5], "col3": [5, 6]})
+dummy_dataframe = pytest.fixture(dummy_dataframe)
 
 
 @pytest.fixture(params=["dummy_dataframe", "dummy_numpy_array"])
